requirements.txt
acme==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certbot==5.3.1
certbot-apache==1.21.0
certifi==2020.6.20
cffi==2.0.0
chardet==4.0.0
click==8.0.3
cloud-init==25.3
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.7.1
configobj==5.0.6
constantly==15.1.0
cryptography==46.0.5
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.133.1
gpg==1.16.0
greenlet==1.1.2
h11==0.16.0
httplib2==0.20.2
httptools==0.7.1
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.0.3
josepy==2.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
msgpack==1.0.3
mysql-connector-python==9.6.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==21.3
parsedatetime==2.6
passlib==1.7.4
pexpect==4.8.0
ptyprocess==0.7.0
pyasn1==0.6.2
pyasn1-modules==0.2.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyOpenSSL==25.3.0
pyparsing==2.4.7
pyRFC3339==2.1.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4.1
python-augeas==0.5.0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-jose==3.5.0
python-magic==0.4.24
python-multipart==0.0.22
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.9.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.9.2
SQLAlchemy==2.0.47
ssh-import-id==5.11
starlette==0.52.1
systemd-python==234
Twisted==22.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-pro-client==8001
ufw==0.36.1
urllib3==1.26.5
uvicorn==0.41.0
uvloop==0.22.1
wadllib==1.3.6
watchfiles==1.1.1
websockets==16.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
ldap3
